+++ hooks/__tests__/useCardCompanySelect.test.ts
@@ -1,7 +1,6 @@
-import { renderHook } from '@testing-library/react';
-import { act } from 'react';
-import { useCardCompanySelect } from '../lib/useCardCompanySelect';
-import { CardCompany } from '../lib/types/card';
+import { renderHook, act } from '@testing-library/react';
+import { useCardCompanySelect } from '../src/lib/useCardCompanySelect';
+import { CardCompany } from '../src/lib/types/card';
 
 describe('useCardCompanySelect custom hook 테스트', () => {
   test.each(['BC카드', '신한카드', '카카오뱅크', '현대카드', '우리카드', '롯데카드', '하나카드', '국민카드'])(
