+++ hooks/__tests__/useCardExpirationInput.test.ts
@@ -1,5 +1,5 @@
 import { renderHook, act } from '@testing-library/react';
-import { useCardExpirationInput } from '../lib/useCardExpirationInput';
+import { useCardExpirationInput } from '../src/lib/useCardExpirationInput';
 
 describe('useCardExpirationInput custom hook 테스트', () => {
   it('카드 유효기간을 입력하면, 에러가 발생하지 않는다.', () => {
